netcfg: Fix type inconsistency in NetCfgChangeType
The D-Bus API used uint32_t while most places in the C++ code used
uint16_t.  This makes no sense.

Since the D-Bus API is external and has been exposed like this for a
long time, that will be preserved.  Switching from uint16_t to uint32_t
gives us just more space to grow even more notification types in the
future.

+++ src/netcfg/netcfg-changeevent.hpp
@@ -81,14 +81,14 @@ struct NetCfgChangeEvent
     }
 
 
-    static const std::vector<std::string> FilterMaskList(const uint16_t mask,
+    static const std::vector<std::string> FilterMaskList(const uint32_t mask,
                                                          bool tech_form = false)
     {
         std::vector<std::string> ret;
 
-        for (uint16_t i = 0; i < 16; ++i)
+        for (uint32_t i = 0; i < 16; ++i)
         {
-            uint16_t flag = 1 << i;
+            uint32_t flag = 1 << i;
             NetCfgChangeType t = (NetCfgChangeType)(flag);
             if (flag & mask)
             {
